+++ hack/urllint/config.json
@@ -13,8 +13,11 @@
         "https://127.0.0.1",
         "http://0.0.0.0",
         "https://0.0.0.0",
+        "https://vault.example.com:8200",
+        "https://github.com/vmware-tanzu",
         "$",
         "<",
-        ">"
+        ">",
+        "@"
     ]
 }
