+++ hack/urllint/main.go
@@ -3,8 +3,6 @@ package main
 import (
 	"flag"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
 
 	urllint "github.com/vmware-tanzu/tce/hack/urllint/pkg/lint"
@@ -16,13 +14,9 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-
 	var pathFlag = flag.String("path", wd, "path to be provided")
-
 	flag.Parse()
-
 	llint, err := urllint.New("config.json")
-
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -30,28 +24,6 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-
 	log.Println("The following is the path that lint is working on: ", *pathFlag)
-
-	for _, link := range llint.LinkLints {
-		if !IsUrl(link.Line) {
-			log.Fatalf("file:%s line:%d:%d Link:%s is invalid link", link.Path, link.Position.Row, link.Position.Col, link.Line)
-			//log.Printf("file:%s line:%d:%d Link:%s has error", link.Path, link.Position.Row, link.Position.Col, link.Line)
-		}
-		resp, err := http.Get(link.Line)
-		if err != nil {
-			log.Fatalf("file:%s line:%d:%d Link:%s is not working.", link.Path, link.Position.Row, link.Position.Col, link.Line)
-			//log.Printf("file:%s line:%d:%d Link:%s has error", link.Path, link.Position.Row, link.Position.Col, link.Line)
-		}
-		if resp.StatusCode >= 300 {
-			log.Fatalf("file:%s line:%d:%d Link:%s returns status code %d", link.Path, link.Position.Row, link.Position.Col, link.Line, resp.StatusCode)
-			//log.Printf("file:%s line:%d:%d Link:%s has error", link.Path, link.Position.Row, link.Position.Col, link.Line)
-		}
-		//fmt.Println(link.Path, ":", link.Line)
-	}
-}
-
-func IsUrl(str string) bool {
-	u, err := url.Parse(str)
-	return err == nil && u.Scheme != "" && u.Host != ""
+	llint.LintAll()
 }
